requirement.txt
sqlalchemy
bcrypt==3.2.2 
passlib==1.7.4
python-dotenv
email-validator==2.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
pymysql 
fastapi[standard]  
jose
bleach

